rotate and bicubic

+++ 電腦圖學導論/正課/Project1 FLTK/Skeleton/Codes/TargaImage.h
@@ -65,7 +65,7 @@ class TargaImage
 
         bool Half_Size();
         bool Double_Size();
-        bool Resize(float scale);
+        bool Resize(double scale);
         bool Rotate(double angleDegrees);
 
     private:
